Move aedt_versions
