+++ src/ansys/aedt/core/visualization/advanced/touchstone_parser.py
@@ -31,9 +31,9 @@ import warnings
 
 from ansys.aedt.core import Edb
 from ansys.aedt.core.aedt_logger import pyaedt_logger as logger
-from ansys.aedt.core.generic.aedt_versions import aedt_versions
 from ansys.aedt.core.generic.file_utils import open_file
 from ansys.aedt.core.generic.general_methods import pyaedt_function_handler
+from ansys.aedt.core.internal.aedt_versions import aedt_versions
 
 try:
     import numpy as np
