+++ src/ansys/aedt/core/visualization/post/spisim.py
@@ -29,7 +29,6 @@ import re
 from struct import unpack
 import subprocess  # nosec
 
-from ansys.aedt.core.generic.aedt_versions import aedt_versions
 from ansys.aedt.core.generic.file_utils import generate_unique_folder_name
 from ansys.aedt.core.generic.file_utils import generate_unique_name
 from ansys.aedt.core.generic.file_utils import open_file
@@ -37,6 +36,7 @@ from ansys.aedt.core.generic.general_methods import env_value
 from ansys.aedt.core.generic.general_methods import pyaedt_function_handler
 from ansys.aedt.core.generic.settings import is_linux
 from ansys.aedt.core.generic.settings import settings
+from ansys.aedt.core.internal.aedt_versions import aedt_versions
 from ansys.aedt.core.visualization.post.spisim_com_configuration_files.com_parameters import COMParametersVer3p4
 from numpy import float64
 from numpy import zeros
